+++ src/mame/sega/coolridr.cpp
@@ -2632,7 +2632,7 @@ void coolridr_state::unk_blit_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 	{
 		default:
 		{
-			printf("sysh1_unk_blit_w unhandled offset %04x %08x %08x\n", offset, data, mem_mask);
+			printf("sysh1_unk_blit_w unhandled offset %04x %08x %08x\n", (uint16_t)offset, data, mem_mask);
 		}
 		break;
 
@@ -2859,7 +2859,7 @@ uint32_t coolridr_state::sound_dma_r(offs_t offset)
 	if(offset == 2 || offset == 6) // DMA status
 		return 0;
 
-	printf("%08x\n",offset);
+	printf("%08x\n", (uint32_t)offset);
 
 	return m_sound_dma[offset];
 }
